# [ci] Speed up a test case