@johnkerl 2025-02-14
@jp-dark -- I bisected the increased runtime to #3647.

This is not really a performance regression per se:

* Before #3647 a `ManagedQuery` object was attached to a `SOMAArray`
* After #3647 a new `ManagedQuery` object is allocated on every `write`
* The case  of 10,000 rows and a cap-nbytes of 100 was a highly artificial setup that created 3900 fragments
* For each of those 3900 fragments, instantiation of the `ManagedQuery` object for that write does a core array open and those core array opens take longer and longer the more fragments there are

Everything is as expected here, and the case of a 10,000 row table being split into 3900 separate writes was a highly artificial setup that is not expected to be indicative of any real-world use of this package.